requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
openai>=2.26.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
pypdf>=5.0.0
faiss-cpu>=1.7.4
reportlab>=4.2.0
